switch to vite; refactor components
