+++ stream_pose_ml/api/app.py
@@ -1,4 +1,5 @@
 import time
+import os
 
 from flask import Flask, request
 from flask_cors import CORS
