+++ stream_pose_ml/run.py
@@ -13,4 +13,4 @@ if __name__ == "__main__":
         
         """
     )
-    socketio.run(app, host="0.0.0.0", port=5000, debug=True, use_reloader=False)
+    socketio.run(app, host="0.0.0.0", port=5001, debug=True, use_reloader=False)
