+++ web_ui/.env
@@ -1,4 +1,4 @@
-REACT_APP_STREAM_POSE_ML_API_ENDPOINT=http://localhost:5001
-REACT_APP_BLUETOOTH_DEVICE_SERVICE_UUID=0000FFE0-0000-1000-8000-00805F9B34FB
-REACT_APP_BLUETOOTH_DEVICE_CHARACTERISTIC_UUID=0000FFE1-0000-1000-8000-00805F9B34FB
-REACT_APP_BLUETOOTH_DEVICE_NAME=HMSoft
+VITE_STREAM_POSE_ML_API_ENDPOINT=http://localhost:5001
+VITE_BLUETOOTH_DEVICE_SERVICE_UUID=0000FFE0-0000-1000-8000-00805F9B34FB
+VITE_BLUETOOTH_DEVICE_CHARACTERISTIC_UUID=0000FFE1-0000-1000-8000-00805F9B34FB
+VITE_BLUETOOTH_DEVICE_NAME=HMSoft
