+++ web_ui/package.json
@@ -2,20 +2,25 @@
   "name": "web_ui",
   "version": "0.1.0",
   "private": true,
+  "type": "module",
   "dependencies": {
     "@testing-library/jest-dom": "^5.16.5",
     "@testing-library/react": "^13.4.0",
     "@testing-library/user-event": "^13.5.0",
+    "@vitejs/plugin-react-swc": "^3.6.0",
     "react": "^18.2.0",
     "react-dom": "^18.2.0",
-    "react-scripts": "5.0.1",
+    "vite": "^5.2.6",
+    "vite-plugin-svgr": "^4.2.0",
+    "vite-tsconfig-paths": "^4.3.2",
     "web-vitals": "^2.1.4"
   },
   "scripts": {
-    "start": "react-scripts start",
-    "build": "react-scripts build",
-    "test": "react-scripts test",
-    "eject": "react-scripts eject"
+    "start": "vite",
+    "build": "tsc && vite build",
+    "serve": "vite preview",
+    "test": "vitest",
+    "test:coverage": "vitest run --coverage --watch=false"
   },
   "eslintConfig": {
     "extends": [
@@ -38,8 +43,11 @@
   "devDependencies": {
     "@mediapipe/pose": "^0.5.1675469404",
     "@mediapipe/tasks-vision": "^0.10.12",
+    "@vitest/coverage-v8": "^1.4.0",
     "axios": "^1.3.4",
     "dotenv": "^16.0.3",
-    "socket.io-client": "^4.6.1"
+    "jsdom": "^24.0.0",
+    "socket.io-client": "^4.6.1",
+    "vitest": "^1.4.0"
   }
 }
